+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Change: Add ID of the workflow run to the output
 - Feature: Support restricting workflow runs examine since a given date
 - Feature: Fetch all the workflow runs, not just the first 100
 - Breaking: Removed the `list-workflows` command and renamed the `duration-history` to just `duration`